// src/Models/Contract/FinanceModel.ts
export type FinanceModel = {
  barcode: string;
  product_name: string;
  orders_count: number;
  sales_count: number;
  redemption_percentage: number;
  prime_cost: number;
  margin_percentage: number;
  store_commission: number;
  logistics_cost: number;
  storage_cost: number;
  tax: number;
  gross_profit: number;
  net_profit: number;
};